refactor(buildings): split PanelComplex and share triangle helpers
Extract `triangle_normal` / `dihedral_kink` into building-components panels.
Split `panel_complex` into types/adjacency/mutate/query/present/parse modules
with category-scoped `impl PanelComplex`. Add compact `FromStr` mesh syntax and
wire `/show panel-complex` in the buildings playground.

## Agent Dialogue

- **Developer:** Deduplicate triangle_normal; split panel_complex into submodules;
  wire playground with FromStr like `1=(…),2=(…) ... {1,2,4},{4,3,5}`.
- **Agent:** Added panels/triangle.rs helpers; modular PanelComplex + put_point;
  parse + playground `--mesh` defaulting to a mild trapezoid.
