+++ maybraid/richmond/buildings-playground/README.md
@@ -26,6 +26,10 @@ cargo run -p richmond-buildings-playground -- show tessellated-triangle --a 0,0 
 cargo run -p richmond-buildings-playground -- show quad-panel
 # Coplanar quad (no joint):
 cargo run -p richmond-buildings-playground -- show quad-panel --a0 0,0,0 --a1 3,0,0 --b0 0,0,3 --b1 3,0,3
+# Panel complex (default mild trapezoid); compact mesh syntax:
+cargo run -p richmond-buildings-playground -- show panel-complex
+cargo run -p richmond-buildings-playground -- show panel-complex \
+  --mesh '1=(0,0,0),2=(1,0,0),3=(0,1,0),4=(0,0,1) ... {1,2,4},{1,4,3}'
 ```
 
 In-game: press `/` for the command console (same clap commands as argv).
@@ -37,6 +41,7 @@ In-game: press `/` for the command console (same clap commands as argv).
 - `show pitch [--rise R] [--run R] [--length L] [--tile-width W] [--left B] [--right B] | [--eave E --ridge R]` — shepherd's-thatch pitched face; omit `--length`/`--left`/`--right` for optional regions; `--eave`+`--ridge` uses equal end triangles
 - `show tessellated-triangle [--a X,Z] [--b X,Z] [--c X,Z]` — rough-stone floor fill of a panel-space triangle
 - `show quad-panel [--a0|--a1|--b0|--b1 X,Y,Z] [--t-a0|--t-a1|--t-b0|--t-b1 T] [--min-dihedral R] [--no-joint]` — two lines → two tessellated triangles + optional crease `JointNode` (default corners are a ~90° fold; thicknesses default to 0.4)
+- `show panel-complex [--mesh 'id=(x,y,z) … {a,b,c}'] [--min-dihedral R] [--no-joint]` — point-id triangle mesh + crease joints; optional thickness as 4th tuple component
 - `show wizards-tower [--noise 0.5]` — authored tower hierarchy (`LodScene` composition)
 - `show stacked-rings [--floor-count N] [--floor-height H] [--radius R]` — circular wall stack (kit scale check)
 - `show bedroom [--extent X,Y,Z] [--noise 0.5] [--spaciousness 1.0] [--occupancy 0.55] [--door]` — hierarchical bedroom; bed-first multi-fill under spaciousness/occupancy; `--door` adds a −Z circulation exclusion
