+++ maybraid/richmond/buildings-playground/src/preview.rs
@@ -16,6 +16,7 @@ use richmond_building_components::roofs::{Pitch, RoofGeometry, RoofNode};
 use richmond_building_components::Placement;
 use richmond_building_components::ComponentsOnly;
 use richmond_buildings::bedroom::Bedroom;
+use richmond_buildings::panel_complex::{PanelComplex, PanelComplexJointPolicy};
 use richmond_buildings::quad_panel::{QuadCorner, QuadPanel, QuadPanelJointPolicy};
 use richmond_buildings::stacked_rings::StackedRings;
 use richmond_buildings::tessellated_triangle_panel::TessellatedTrianglePanel;
@@ -65,6 +66,11 @@ pub enum PreviewSubject {
 		t_a1: f32,
 		t_b0: f32,
 		t_b1: f32,
+		min_dihedral: f32,
+		no_joint: bool,
+	},
+	PanelComplex {
+		mesh: String,
 		min_dihedral: f32,
 		no_joint: bool,
 	},
@@ -154,6 +160,13 @@ impl PreviewConfig {
 				no_joint,
 			} => format!(
 				"preview: quad-panel (a0={a0:?} a1={a1:?} b0={b0:?} b1={b1:?} t=[{t_a0:.2},{t_a1:.2},{t_b0:.2},{t_b1:.2}] min_dihedral={min_dihedral:.3} no_joint={no_joint})"
+			),
+			PreviewSubject::PanelComplex {
+				ref mesh,
+				min_dihedral,
+				no_joint,
+			} => format!(
+				"preview: panel-complex (mesh={mesh:?} min_dihedral={min_dihedral:.3} no_joint={no_joint})"
 			),
 			PreviewSubject::Polyline => "preview: partition polyline (L)".into(),
 			PreviewSubject::LinearWall => "preview: walling linear-wall (door)".into(),
@@ -452,6 +465,30 @@ pub fn present_preview_lod(
 				transform,
 				ComponentsOnly(panel).scene_with_lod(&lod_ref),
 			);
+		}
+		PreviewSubject::PanelComplex {
+			mesh,
+			min_dihedral,
+			no_joint,
+		} => {
+			let policy = if *no_joint {
+				PanelComplexJointPolicy::never()
+			} else {
+				PanelComplexJointPolicy::min_dihedral_rad(*min_dihedral)
+			};
+			match mesh.parse::<PanelComplex>() {
+				Ok(complex) => {
+					let complex = complex.with_joint_policy(policy);
+					spawn_preview(
+						&mut commands,
+						transform,
+						ComponentsOnly(complex).scene_with_lod(&lod_ref),
+					);
+				}
+				Err(e) => {
+					warn!("panel-complex parse failed: {e}");
+				}
+			}
 		}
 		PreviewSubject::Polyline => {
 			let node = PartitionNode::rough_stone(
